Uses the correct repo path as argument to svnadmin dump
The old one would have been something along the lines of
`/var/define/templates/vc-repos/svn/[project_name]`.

When a path to a non-directory is given to `svnadmin dump` it writes
an error to stderr and nothing to stdout.  The error was logged,
but stdout was redirected to a file anyway which `svnadmin load`
would fail to read when creating a new project using the template.

Fixes #4191

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@23819 484df14f-75d2-4b3f-8ba0-490e05069a01